some changing

+++ Headers/binTranslator.h
@@ -5,7 +5,7 @@
 
 struct BinTranslator
 {
-	void numToBin(int dec_num);
+	void numToBin(float dec_num);
 
 	int bin_num = 0;
 	std::string str_bin_num = "";
